clean cairocoder
