+++ plugins/cairocoder/src/actions/fixCairoCode.ts
@@ -1,9 +1,9 @@
 import { StarknetAgentInterface } from '@starknet-agent-kit/agents';
-import { z } from 'zod';
 import { fixCairoCodeSchema } from '../schema/schema.js';
 import { callCairoGenerationAPI, extractCairoCode, saveToDebugFile } from '../utils/utils.js';
 import { addProgram } from '../utils/db_add.js';
 import { retrieveProjectData } from '../utils/db_init.js';  
+import { z } from 'zod';
 
 /**
  * Fix Cairo code using AI via API and update it in the database
@@ -19,7 +19,6 @@ export const fixCairoCode = async (
     console.log('\n➜ Fixing Cairo code');
     console.log(JSON.stringify(params, null, 2));
 
-    // Validate parameters
     if (!params?.programName || !params.programName.endsWith('.cairo')) {
       throw new Error('Program name is required and must end with .cairo');
     }
@@ -28,7 +27,6 @@ export const fixCairoCode = async (
       throw new Error('Error description is required for fixing Cairo code');
     }
 
-    // Retrieve the program from the database
     const projectData = await retrieveProjectData(agent, params.projectName);
     const program = projectData.programs.find(p => p.name === params.programName);
     
@@ -36,13 +34,11 @@ export const fixCairoCode = async (
       throw new Error(`Program ${params.programName} not found in the database`);
     }
     
-    // Extraire l'erreur exacte
     let errorText = params.error;
     if (errorText.includes('[EXACT_ERROR_BEGIN]') && errorText.includes('[EXACT_ERROR_END]')) {
       errorText = errorText.split('[EXACT_ERROR_BEGIN]')[1].split('[EXACT_ERROR_END]')[0].trim();
     }
     
-    // Create a prompt for fixing the code
     const fixPrompt = `I have the following Cairo code that has an error:
 \`\`\`cairo
 ${program.source_code}
@@ -52,18 +48,13 @@ The error is: ${errorText}
 
 Can you fix the compilation errors?`;
     
-    // Generate fixed code
     const generatedContent = await callCairoGenerationAPI(fixPrompt);
     const fixedCairoCode = extractCairoCode(generatedContent);
     
-    // console.log("\nFixed Cairo code = ", fixedCairoCode);
-    
-    // Save to debug file
     const debugFile = saveToDebugFile(params.programName, fixedCairoCode);
     
     await addProgram(agent, projectData.id, params.programName, fixedCairoCode);
     
-    // Return success response
     return JSON.stringify({
       status: 'success',
       message: `Cairo code fixed and updated in database as ${params.programName}`,
