+++ plugins/cairocoder/src/actions/generateCairoCode.ts
@@ -23,7 +23,6 @@ export const generateCairoCode = async (
     
     const generatedContent = await callCairoGenerationAPI(params.prompt);
     const cairoCode = extractCairoCode(generatedContent);
-    // console.log("\nCairo code = ", cairoCode);
     
     const debugFile = saveToDebugFile(params.programName, cairoCode);
