+++ plugins/cairocoder/src/actions/registerProject.ts
@@ -62,9 +62,6 @@ export const registerProject = async (
 
     
     const projectData = await retrieveProjectData(agent, params.projectName);
-    // console.log(
-    //   `Project ${params.projectName}: ${alreadyRegistered ? 'updated' : 'created'}`
-    // );
 
     return JSON.stringify({
       status: 'success',
